requirements.txt
requests
PyYaml
pylint
flake8
